paper_trading/requirements.txt
# Core Dependencies
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0
gymnasium>=0.28.0

# Data and Market Access
yfinance>=0.2.0
requests>=2.25.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=0.19.0

# ElegantRL (already installed in parent directory)
# elegantrl

# Optional: Additional data providers
# alpaca-trade-api>=2.0.0
# polygon-api-client>=1.0.0

# Development and Testing
pytest>=6.0
black>=21.0
flake8>=3.8 